Enforce Record Type Management user permissions
- also add empty table message

+++ app/Http/Controllers/RecordTypeController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\RecordType;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Gate;
 
 class RecordTypeController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-record-types');
+
         $record_types = RecordType::all();
         $title = "Record Types";
         return view('record_types', compact('record_types', 'title'));
@@ -17,6 +20,8 @@ class RecordTypeController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create-record-types');
+
         RecordType::create($request->validate([
             'name' => 'required|unique:record_types,name|max:255',
             'description' => 'nullable|string',
@@ -27,6 +32,8 @@ class RecordTypeController extends Controller
 
     public function update(Request $request, RecordType $record_type)
     {
+        Gate::authorize('edit-record-types');
+
         $record_type->update($request->validate([
             'name' => 'required|max:255|unique:record_types,name,' . $record_type->id,
             'description' => 'nullable|string',
@@ -37,6 +44,8 @@ class RecordTypeController extends Controller
 
     public function destroy(RecordType $record_type)
     {
+        Gate::authorize('delete-record-types');
+        
         try {
             $record_type->delete();
             return back()->with('success', 'Record Type Deleted Successfully!');
